backend/src/requirements.txt
python-twitter
requests
python-dotenv
tweepy
tweepy[async]